This merge commit includes 2 merge conflicts:

* openvidu-browser/src/OpenVidu/Session.ts
In modern code, the OpenVidu version is obtained through other means.
However in the mediasoup branch this value is used in several of the
*LEGACY.ts files.

* openvidu-browser/src/OpenViduInternal/WebRtcPeer/WebRtcPeer.ts
In the mediasoup branch, the special treatment for ionic was removed:
    openvidu-browser: removed Ionic iOS timeout on first subscription

https://github.com/OpenVidu/openvidu/commit/23d64be8063f8fdb2a212ca845e304762f2803f5
and also the method was converted into async and returned a Promise.
However in master the code is still like the old one.

+++ openvidu-server/docker/openvidu-server-kms/create_image.sh
@@ -1,5 +1,8 @@
-cp ../../target/openvidu-server-"$1".jar ./openvidu-server.jar
-
-docker build -t openvidu/openvidu-server-kms .
-
-rm ./openvidu-server.jar
+VERSION=$1
+if [[ ! -z $VERSION ]]; then
+    cp ../../target/openvidu-server-*.jar ./openvidu-server.jar
+    docker build -t openvidu/openvidu-server-kms:$VERSION .
+    rm ./openvidu-server.jar
+else
+    echo "Error: You need to specify a version as first argument"
+fi
